10229: remove obsolete getCasesAssociatedWithUser, replace with identical getCaseForUser
